[NAYB-94] refactor : ItemFixture 패키지 경로 변경
